# ingest_openapi.yaml
openapi: 3.0.0
info:
  version: v1.0.0
  title: ''
  description: ''
paths:
  /service-info:
    get:
      description: Returns information about the ingest service
      operationId: ingest_operations.get_service_info
      responses:
        200:
          description: Retrieve info about the ingest service
          content:
            application/json:
              schema:
                type: object
  /add-s3-credential:
    post:
      description: Add credentials for an S3 bucket
      operationId: ingest_operations.add_s3_credential
      requestBody:
        $ref: '#/components/requestBodies/S3CredentialRequest'
      responses:
          200:
            description: Success
            content:
              application/json:
                schema:
                  type: object
  /ingest/moh_variants/{program_id}:
    post:
      description: Add a genomic file that corresponds to a SampleRegistration in an MoH Program schema.
      operationId: ingest_operations.add_moh_variant
      parameters:
        - $ref: "#/components/parameters/ProgramId"
      requestBody:
        $ref: '#/components/requestBodies/MoHVariantRequest'
      responses:
          200:
            description: Success
            content:
              application/json:
                schema:
                  type: object
  /ingest/clinical_donors:
    post:
      description: Add a list of donors with clinical data produced by the clinical ETL.
      operationId: ingest_operations.add_clinical_donors
      requestBody:
        $ref: "#/components/requestBodies/ClinicalDonorRequest"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
components:
  parameters:
    ProgramId:
      in: path
      name: program_id
      required: true
      description: name of an MoH program to ingest this genomic file into
      schema:
        type: string
  requestBodies:
    S3CredentialRequest:
      content:
        'application/json':
          schema:
            type: object
            properties:
              endpoint:
                type: string
                description: URL to the endpoint
                example: http://candig.docker.internal:9000
              bucket:
                type: string
                description: name of the bucket
              access_key:
                type: string
                description: access key for the bucket
              secret_key:
                type: string
                description: secret key for the bucket
            required:
              - endpoint
              - bucket
              - access_key
              - secret_key
    MoHVariantRequest:
      content:
        'application/json':
          schema:
            type: object
            properties:
              genomic_id:
                type: string
                description: the name of the VCF file, including gz, if present
                example: "HG00096.vcf.gz"
              access_method:
                type: string
                anyOf:
                  - $ref: "#/components/schemas/S3Access"
                  - $ref: "#/components/schemas/FileAccess"
              samples:
                type: array
                items:
                  $ref: "#/components/schemas/VCFSample"
            required:
              - genomic_id
              - access_method
              - samples
    ClinicalDonorRequest:
      content:
        'application/json':
          schema:
            $ref: "#/components/schemas/DonorWithClinicalDataField"
  schemas:
    S3Access:
      type: string
      description: a specification for an S3 bucket
      pattern: s3:\/\/(.+)\/(.+)
      example: s3://http://candig.docker.internal:9000/mohccn-data
    DonorWithClinicalDataField: # "Sanity check" validation, everyting else happens in ETL validation/katsu
      type: array
      description: A DonorWithClinicalData Field (e.g. Donor, Biomarker, etc)
      items:
        properties:
          primary_diagnoses:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          treatments:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          specimens:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          exposures:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          radiation:
            type: object
          surgeries:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          followups:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          comorbidities:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          hormone_therapies:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          immunotherapies:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          sample_registrations:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          biomarkers:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          chemotherapies:
            $ref: '#/components/schemas/DonorWithClinicalDataField'
          surgery:
            type: object
        additionalProperties:
          oneOf:
            - type: string
              nullable: true
            - type: number
            - type: boolean
            - type: array
              items:
                oneOf:
                  - type: string
                    nullable: true
                  - type: number
                  - type: boolean
                  - type: array
    FileAccess:
      type: string
      description: a file path to a directory on the HTSGet server itself
      pattern: file:\/\/\/(.+)
      example: file://data/samples/
    VCFSample:
      type: object
      properties:
        sample_registration_id:
          type: string
          description: a SampleRegistrationId in the MoH program
          example: sample_registration_id_1
        sample_name_in_file:
          type: string
          description: the name of the sample in the VCF file
          example: TUMOUR/NORMAL/PROGRAM_SAMPLE_REGISTRATION_ID_1/...
      required:
        - sample_registration_id
        - sample_name_in_file